renamed price to pricefetch
