+++ fetchprice/go.mod
@@ -1,0 +1,3 @@
+module github.com/cocobokostudios/vgl-tools/fetchprice
+
+go 1.18
