+++ fetchprice/lib/fetchprice_test.go
