--- price/go.mod
@@ -1,3 +1,0 @@
-module github.com/cocobokostudios/vgl-tools/price
-
-go 1.18
